requirements.txt
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128

#torch>=1.9.0
#torchvision>=0.10.0
numpy>=1.19.5
opencv-python>=4.5.3
pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.2
pycocotools>=2.0.2 